Refactor class names and improve documentation
