+++ misc/semver-checker/modules/semver-checker-core/src/main/java/io/ballerina/semver/checker/diff/NodeDiff.java
@@ -20,130 +20,15 @@ package io.ballerina.semver.checker.diff;
 
 import io.ballerina.compiler.syntax.tree.Node;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Optional;
+/**
+ * Base representation of changes in Ballerina syntax tree nodes.
+ *
+ * @param <T> node type
+ * @since 2201.2.0
+ */
+public interface NodeDiff<T extends Node> extends Diff {
 
-public class NodeDiff<T extends Node> implements INodeDiff<T> {
+    T getNewNode();
 
-    protected final T newNode;
-    protected final T oldNode;
-    protected DiffType diffType;
-    protected CompatibilityLevel compatibilityLevel;
-    protected final List<IDiff> childDiffs;
-    protected String message;
-
-    public NodeDiff(T newNode, T oldNode) {
-        this(newNode, oldNode, CompatibilityLevel.UNKNOWN);
-    }
-
-    public NodeDiff(T newNode, T oldNode, CompatibilityLevel compatibilityLevel) {
-        this.newNode = newNode;
-        this.oldNode = oldNode;
-        this.compatibilityLevel = compatibilityLevel;
-        this.childDiffs = new ArrayList<>();
-        this.message = null;
-
-        if (newNode != null && oldNode == null) {
-            this.diffType = DiffType.NEW;
-        } else if (newNode == null && oldNode != null) {
-            this.diffType = DiffType.REMOVED;
-        } else if (newNode != null) {
-            this.diffType = DiffType.MODIFIED;
-        } else {
-            this.diffType = DiffType.UNKNOWN;
-        }
-    }
-
-    @Override
-    public T getNewNode() {
-        return newNode;
-    }
-
-    @Override
-    public T getOldNode() {
-        return oldNode;
-    }
-
-    @Override
-    public DiffType getType() {
-        return diffType;
-    }
-
-    @Override
-    public void setType(DiffType diffType) {
-        this.diffType = diffType;
-    }
-
-    @Override
-    public CompatibilityLevel getCompatibilityLevel() {
-        return compatibilityLevel;
-    }
-
-    @Override
-    public void setCompatibilityLevel(CompatibilityLevel compatibilityLevel) {
-        this.compatibilityLevel = compatibilityLevel;
-    }
-
-    @Override
-    public void computeCompatibilityLevel() {
-        if (compatibilityLevel == CompatibilityLevel.UNKNOWN) {
-            compatibilityLevel = childDiffs.stream()
-                    .map(IDiff::getCompatibilityLevel)
-                    .max(Comparator.comparingInt(CompatibilityLevel::getRank))
-                    .orElse(CompatibilityLevel.UNKNOWN);
-        }
-    }
-
-    public List<IDiff> getChildDiffs() {
-        return Collections.unmodifiableList(childDiffs);
-    }
-
-    public Optional<String> getMessage() {
-        return Optional.ofNullable(message);
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public void addChildDiff(IDiff childDiff) {
-        this.childDiffs.add(childDiff);
-    }
-
-    public void addChildDiffs(List<? extends IDiff> childDiffs) {
-        this.childDiffs.addAll(childDiffs);
-    }
-
-    /**
-     * Derives the parent {@link CompatibilityLevel} of a given source (syntax node) based on the compatibility
-     * information of its child elements.
-     *
-     * @param childCompatibilities compatibility information of the child elements
-     * @return parent {@link CompatibilityLevel}
-     */
-    protected static CompatibilityLevel getUnifiedCompatibility(CompatibilityLevel... childCompatibilities) {
-        return Arrays.stream(childCompatibilities)
-                .max(Comparator.comparingInt(CompatibilityLevel::getRank))
-                .orElse(CompatibilityLevel.UNKNOWN);
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        if (childDiffs == null || childDiffs.isEmpty()) {
-            sb.append("compatibility: ").append(compatibilityLevel.toString()).append(", ");
-            if (message != null) {
-                sb.append("description: ").append(message).append(System.lineSeparator());
-                return sb.toString();
-            }
-        } else {
-            childDiffs.forEach(diff -> sb.append(diff.toString()));
-        }
-
-        return sb.toString();
-    }
+    T getOldNode();
 }
