+++ misc/semver-checker/modules/semver-checker-core/src/main/java/io/ballerina/semver/checker/util/PackageUtils.java
@@ -34,6 +34,8 @@ import java.util.Map;
 
 public class PackageUtils {
 
+    public static final String QUALIFIER_PUBLIC = "public";
+
     /**
      * Loads the target ballerina source project instance using the Project API, from the file path of the open/active
      * editor instance in the client(plugin) side.
